src/django_server/requirements.txt
aniso8601==3.0.2
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-cors-headers==2.4.0
django-filter==2.1.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
httpie==1.0.2
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
oauthlib==3.0.0
Pillow==5.4.1
promise==2.2.1
Pygments==2.3.0
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.0.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
